geoJson Project/backendCode/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions

brotli==1.0.9
ConfigParser==5.2.0
cryptography==3.4.7
Cython==0.29.27
dl==0.1.0
docutils==0.18.1
HTMLParser==0.0.2
importlib_metadata==4.10.1
ipaddr==2.2.0
keyring==23.5.0
lockfile==0.12.2
lxml==4.6.3
matplotlib==3.4.1
mock==4.0.3
mypy_extensions==0.4.3
numpy==1.19.5
ordereddict==1.1
protobuf==3.19.4
pyOpenSSL==22.0.0
simplejson==3.17.6
typing_extensions==3.7.4.3
unicodedata2==14.0.0